# ISI TP2/swagger.json
{
  "swagger": "2.0",
  "info": {
    "name": "IPBOOK",
    "version": "1.0.0"
  },
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/livro/all": {
      "get": {
         "summary": "Todos os livros",
         "tags":["Livro"],
         "parameters": [],
        "responses": {
           "200": {
             "description": "A lista dos livros",
             "schema": {
               "type": "array",
               "items": {
                "$ref": "#/definitions/livros"
              }
            }
           },
          "400": {
              "description": "Algo correu mal"
          },
          "500": {
            "description": "Ocorreu algum erro para obter os dados dos livros"
          }
        }
       }
     },
     "/livro/{bookID}": {
       "get": {
         "summary": "Mostra o livro cuja id selecionas-te",
         "tags":["Livro"],
         "operationId":"getLivroById",
         "parameters": [
          {
            "name": "bookID",
             "in": "path",
             "description": "ID do livro",
            "required": true,
            "type": "integer",
             "format": "int64"
         }
         ],
         "responses": {
           "200": {
             "description": "ID Livro",
             "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/livros"
              }
             }
           },
           "400": {
              "description": "Algo correu mal"
          },
          "500": {
            "description": "Ocorreu um erro a obter informacao sobre o livro"
          }
        }
      }
    },
    "/livro/delete/{bookID}": {
      "delete": {
        "summary": "Remove o livro cuja id selecionas-te",
        "tags":["Livro"],
        "operationId":"DeleteLivro",
        "parameters": [
         {
           "name": "bookID",
            "in": "path",
            "description": "ID do livro",
           "required": true,
           "type": "integer",
            "format": "int64"
        }
        ],
        "responses": {
          "200": {
            "description": "Livro removido com sucesso",
            "schema": {
             "type": "array",
             "items": {
               "$ref": "#/definitions/livros"
             }
            }
          },
          "400": {
             "description": "Algo correu mal"
         },
         "500": {
           "description": "Não foi possivel remover o livro"
         }
       }
     }
   },
   "/livro/add": {
    "post": {
       "summary": "Adiciona um livro",
       "tags":["Livro"],
       "operationId":"AddLivro",
       "parameters": [
        {
           "name": "body",
           "in": "body",
           "description": "Adiciona um livro",
           "required": true,
           "type": "string",
           "schema": {
            "$ref": "#/definitions/livros"
          }
         }
      ],
      "responses": {
         "200": {
           "description": "Livro adicionado com sucesso",
           "schema": {
             "type": "array",
             "items": {
              "$ref": "#/definitions/livros"
            }
          }
         },
        "400": {
            "description": "Campos Não preenchidos"
        },
        "500": {
          "description": "Ocorreu algum erro ao adicionar o livro"
        }
      }
     }
   },
   "/user/all": {
    "get": {
       "summary": "Todos os utilizadores",
       "tags":["Utilizador"],
       "parameters": [],
      "responses": {
         "200": {
           "description": "A lista dos utilizadores",
           "schema": {
             "type": "array",
             "items": {
              "$ref": "#/definitions/users"
            }
          }
         },
        "400": {
            "description": "Algo correu mal"
        },
        "500": {
          "description": "Ocorreu algum erro para obter os dados dos utilizadores"
        }
      }
     }
   },
    "/user/signup": {
      "post": {
        "summary": "Registar Utilizador",
        "tags":["Utilizador"],
        "operationId":"RegisterUtilizador",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Cria um Utilizador",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/definitions/users"
            }
        }
        ],
        "responses": {
          "200": {
            "description": "Utilizador registado",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/users"
              }
            }
          },
          "400": {
            "description": "Conteudo nao pode estar vazio!"
          },
          "500": {
            "description": "Ocorreu algum erro ao criar o Utilizador"
          }
        }
      }
    },
    "/user/update/{id}": {
      "put": {
        "summary": "Atualizar Utilizador",
        "tags":["Utilizador"],
        "operationId":"UpdateUtilizador",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do utilizador",
            "required": true,
            "type": "integer",
            "format": "int32",
            "schema": {
              "$ref": "#/definitions/users"
            }
        }, 
        {
          "name": "body",
            "in": "body",
            "description": "Nome do utilizador",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/definitions/users"
            }
        }
        ],
        "responses": {
          "200": {
            "description": "Utilizador atualizado com sucesso !",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/users"
              }
            }
          },
          "400": {
            "description": "Conteudo nao pode estar vazio!"
          },
          "500": {
            "description": "Ocorreu algum erro ao atualizar o conteudo"
          }
        }
      }
    },
    "/user/delete/{id}": {
      "delete": {
        "summary": "Remove o utilizador cuja id selecionas-te",
        "tags":["Utilizador"],
        "operationId":"DeleteUtilizador",
        "parameters": [
         {
           "name": "id",
            "in": "path",
            "description": "ID do utilizador",
            "required": true,
            "type": "integer",
            "format": "int64"
        }
        ],
        "responses": {
          "200": {
            "description": "Utilizador removido com sucesso",
            "schema": {
             "type": "array",
             "items": {
               "$ref": "#/definitions/users"
             }
            }
          },
          "400": {
             "description": "Algo correu mal"
         },
         "500": {
           "description": "Não foi possivel eliminar o utilizador"
         }
       }
     }
   },
   "/user/login": {
    "post": {
      "summary": "LogIn de um Utilizador",
      "tags":["Utilizador"],
      "operationId":"GetLogin",
      "parameters": [
        {
          "name": "body",
          "in": "body",
          "description": "Faz o login de um Utilizador",
          "required": true,
          "type": "string",
          "schema": {
            "$ref": "#/definitions/users"
          }
      }
      ],
      "responses": {
        "200": {
          "description": "Autenticado com sucesso",
          "schema": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/users"
            }
          }
        },
        "400": {
          "description": "Utilizador não encontrado"
        },
        "500": {
          "description": "Erro ao efetuar o login"
        }
      }
    }
  },
  "/user/{id}": {
    "get": {
      "summary": "Mostra o utilizador cuja id selecionas-te",
      "tags":["Utilizador"],
      "operationId":"GetUtilizadorInfo",
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "description": "ID do utilizador",
          "required": true,
          "type": "integer",
          "format": "int32",
          "schema": {
            "$ref": "#/definitions/users"
          }
        } 
      ],
      "responses": {
        "200": {
          "description": "Autenticado com sucesso",
          "schema": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/users"
            }
          }
        },
        "400": {
          "description": "Utilizador não encontrado"
        },
        "500": {
          "description": "Erro ao efetuar o login"
        }
      }
    }
  }
  },
  "definitions": {
    "livros": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "authors": {
          "type": "string"
        },
        "average_rating": {
          "type": "string"
        },
        "isbn": {
          "type": "string"
        },
        "num_pages": {
          "type": "string"
        },
        "publication_date": {
          "type": "string"
        },
        "publisher": {
          "type": "string"
        }
      }
    },
    "users": {
      "type": "object",
      "properties": {
        "nome": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "nif": {
          "type": "string"
        },
        "TipoUtilizador_id": {
          "type": "integer",
          "format": "int64"
        }  
      }
    }
  }
}